# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi API",
        "description": "Documentación de la API para mi aplicación",
        "contact": {
            "email": "friendhub@contact.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/workouts": {
            "get": {
                "summary": "Obtener todos los entrenamientos",
                "description": "Devuelve una lista de todos los entrenamientos disponibles",
                "operationId": "0d949ad146e8fdb6a4eab0308be812b7",
                "responses": {
                    "200": {
                        "description": "Lista de entrenamientos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crear un nuevo entrenamiento",
                "description": "Crea un nuevo entrenamiento con los ejercicios asignados",
                "operationId": "e2281cab06cbaf5ecdcc53e170217d4b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "exercises"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "exercises": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "exercise_id": {
                                                    "type": "integer"
                                                },
                                                "sets": {
                                                    "type": "integer"
                                                },
                                                "reps": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entrenamiento creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            }
        },
        "/api/workouts/{id}": {
            "get": {
                "summary": "Obtener un entrenamiento por ID",
                "description": "Devuelve los detalles de un entrenamiento específico por su ID",
                "operationId": "e17d8de1e05f71253d2f09937087f992",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del entrenamiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del entrenamiento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entrenamiento no encontrado"
                    }
                }
            },
            "put": {
                "summary": "Actualizar un entrenamiento",
                "description": "Actualiza un entrenamiento y sus ejercicios",
                "operationId": "8c2db466f67f7db14cbc031efd87dace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del entrenamiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "exercises"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "exercises": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "exercise_id": {
                                                    "type": "integer"
                                                },
                                                "sets": {
                                                    "type": "integer"
                                                },
                                                "reps": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entrenamiento actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Entrenamiento no encontrado"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar un entrenamiento",
                "description": "Elimina un entrenamiento específico por ID",
                "operationId": "fdd632fc4252aa89a4527a7d7e36a408",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del entrenamiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entrenamiento eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entrenamiento no encontrado"
                    }
                }
            }
        }
    }
}